More Xcode 7 maintenance

+++ src/extended/SampledInstrumentNode.cpp
@@ -54,7 +54,7 @@ namespace LabSound
 
 		voices.resize(maxVoiceCount);
 
-		for (int i = 0; i < voices.size(); i++)
+		for (size_t i = 0; i < voices.size(); i++)
 		{
 			voices[i] = std::make_shared<AudioBufferSourceNode>(sampleRate);
 			voices[i]->connect(ctx, gainNode.get(), 0, 0);
